feat: complete EventBus migration, demand-driven file downloads
Finalize event system and file download improvements:
- Replace AudioPipelineEvent with EventBus
- Propagate EventBus via StreamType::Events associated type
- Extract FileDownloader into separate module
- Add demand_signal to interrupt blocked downloads
- Cancel sequential download on seek, switch to gap-filling
