+++ crates/kithara-audio/Cargo.toml
@@ -11,6 +11,7 @@ workspace = true
 
 [dependencies]
 kithara-decode = { workspace = true }
+kithara-events = { workspace = true }
 kithara-stream = { workspace = true }
 kithara-storage = { workspace = true }
 kithara-bufpool = { workspace = true }
@@ -64,12 +65,12 @@ unimock = { workspace = true }
 
 [[example]]
 name = "file_audio"
-required-features = ["rodio", "memprof"]
+required-features = ["rodio", "memprof", "perf", "kithara-net/perf"]
 
 [[example]]
 name = "hls_audio"
-required-features = ["rodio", "memprof"]
+required-features = ["rodio", "memprof", "perf", "kithara-net/perf", "kithara-hls/perf"]
 
 [[example]]
 name = "hls_drm_audio"
-required-features = ["rodio", "memprof"]
+required-features = ["rodio", "memprof", "perf", "kithara-net/perf", "kithara-hls/perf"]
