+++ crates/kithara-audio/examples/file_audio.rs
@@ -6,16 +6,15 @@
 
 use std::{env::args, error::Error};
 
-use kithara_audio::{Audio, AudioConfig};
+use kithara_audio::{Audio, AudioConfig, EventBus};
 use kithara_file::{File, FileConfig};
 use kithara_stream::{Stream, ThreadPool};
-use tokio::sync::broadcast;
 use tracing::{info, metadata::LevelFilter, warn};
 use tracing_subscriber::EnvFilter;
 use url::Url;
 
 #[tokio::main(flavor = "current_thread")]
-#[hotpath::main]
+#[cfg_attr(feature = "perf", hotpath::main)]
 async fn main() -> Result<(), Box<dyn Error + Send + Sync>> {
     tracing_subscriber::fmt()
         .with_env_filter(
@@ -40,12 +39,16 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync>> {
 
     info!("Opening file: {url}");
 
-    let (events_tx, mut events_rx) = broadcast::channel(128);
+    let bus = EventBus::new(128);
+    let mut events_rx = bus.subscribe();
     let hint = url.path().rsplit('.').next().map(|ext| ext.to_lowercase());
     let pool = ThreadPool::with_num_threads(2)?;
-    let mut config = AudioConfig::<File>::new(FileConfig::new(url.into()).with_thread_pool(pool))
-        .with_prefer_hardware(true)
-        .with_events(events_tx);
+    let mut config = AudioConfig::<File>::new(
+        FileConfig::new(url.into())
+            .with_thread_pool(pool)
+            .with_events(bus),
+    )
+    .with_prefer_hardware(true);
     if let Some(ext) = hint {
         config = config.with_hint(ext);
     }
@@ -78,7 +81,7 @@ async fn main() -> Result<(), Box<dyn Error + Send + Sync>> {
             recv = events_rx.recv() => {
                 match recv {
                     Ok(ev) => info!(?ev),
-                    Err(broadcast::error::RecvError::Lagged(n)) => warn!(n, "events lagged"),
+                    Err(tokio::sync::broadcast::error::RecvError::Lagged(n)) => warn!(n, "events lagged"),
                     Err(_) => break,
                 }
             }
