+++ crates/kithara-audio/src/lib.rs
@@ -26,13 +26,10 @@
 //!     play_audio(chunk);
 //! }
 //!
-//! // Events
-//! let mut events = audio.events();
+//! // Events via decode_events()
+//! let mut events = audio.decode_events();
 //! while let Ok(event) = events.recv().await {
-//!     match event {
-//!         AudioPipelineEvent::Stream(e) => println!("Stream: {:?}", e),
-//!         AudioPipelineEvent::Audio(e) => println!("Audio: {:?}", e),
-//!     }
+//!     println!("Audio: {:?}", event);
 //! }
 //! ```
 
@@ -40,7 +37,6 @@
 #![cfg_attr(test, allow(clippy::ignored_unit_patterns, clippy::allow_attributes))]
 
 // Internal modules
-mod events;
 mod pipeline;
 mod resampler;
 #[cfg(feature = "rodio")]
@@ -49,7 +45,7 @@ mod traits;
 mod types;
 
 // Public API exports
-pub use events::{AudioEvent, AudioPipelineEvent};
+pub use kithara_events::{AudioEvent, EventBus};
 pub use pipeline::{Audio, AudioConfig};
 pub use resampler::ResamplerQuality;
 #[doc(hidden)]
