+++ crates/kithara-hls/Cargo.toml
@@ -35,7 +35,7 @@ hotpath = { workspace = true, optional = true }
 
 [features]
 default = []
-perf = ["dep:hotpath", "hotpath/hotpath"]
+perf = ["dep:hotpath", "hotpath/hotpath", "kithara-net/perf"]
 
 [dev-dependencies]
 kithara-net = { workspace = true, features = ["test-utils"] }
