+++ crates/kithara/Cargo.toml
@@ -12,6 +12,7 @@ workspace = true
 [dependencies]
 kithara-audio = { workspace = true }
 kithara-decode = { workspace = true }
+kithara-events = { workspace = true }
 kithara-stream = { workspace = true }
 kithara-file = { workspace = true, optional = true }
 kithara-hls = { workspace = true, optional = true }
